agetty: split out issue file functions to issuefile.c
Move issue file parsing, display and reload functions to a new
issuefile.c file: agetty_print_issue_file(), agetty_eval_issue_file(),
agetty_show_issue(), agetty_issue_is_changed(), agetty_reload(),
read_os_release(), output_special_char(), and all issuedir/issuefile
read helpers and network interface display functions.

Keep do_prompt(), get_logname() and wait_for_term_input() in
agetty.c as they are tightly coupled to the main login flow.
